requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
joblib==1.2.0
numpy==1.24.3
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3